diagram-service/pyproject.toml
[tool.poetry]
name = "diagram-service"
version = "0.1.0"
description = "ML-сервис для двухстороннего преобразования диаграмм: изображение ↔ текстовое описание"
authors = ["Your Name <your.email@example.com>"]
readme = "README.md"
packages = [{include = "src"}]

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.109.0"
uvicorn = {extras = ["standard"], version = "^0.27.0"}
pydantic = "^2.5.0"
pydantic-settings = "^2.1.0"
python-multipart = "^0.0.6"
loguru = "^0.7.2"
opencv-python = "^4.9.0"
numpy = "^1.26.0"
pillow = "^10.2.0"
ultralytics = "^8.1.0"
paddleocr = "^2.7.0"
paddlepaddle = "^2.6.0"
networkx = "^3.2"
jinja2 = "^3.1.3"
spacy = "^3.7.0"
natasha = "^1.6.0"
pygraphviz = "^1.11"
python-dotenv = "^1.0.0"
aiofiles = "^23.2.1"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
pytest-asyncio = "^0.21.0"
pytest-cov = "^4.1.0"
httpx = "^0.26.0"
black = "^24.1.0"
ruff = "^0.1.0"
mypy = "^1.8.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 100
target-version = ['py310']
include = '\.pyi?$'

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
